requirements.txt
torch==2.2.2
numpy==1.26.4
PyYAML==6.0.2
scipy==1.11.4
# mujoco optional for later high-fidelity sims
# mujoco==3.1.3
# coremltools optional for export on macOS
# coremltools==7.1
pytest==8.3.1
